floats to double

+++ include/interval.h
@@ -6,13 +6,13 @@ class Interval final
 {
 public:
     Interval() = default;
-    Interval(float t0, float t1);
+    Interval(double t0, double t1);
 
     bool overlap(const Interval& another_interval) const;
 
     bool valid() const;
     void print() const;
 
-    float t0 = NAN;
-    float t1 = NAN;
+    double t0 = NAN;
+    double t1 = NAN;
 };
