+++ unit_tests/test_vector.cpp
@@ -15,9 +15,9 @@ TEST(TestVecPlusCorrectness, Subtest_1)
 
     Vector3D v_sum = v1 + v2;
 
-    ASSERT_NEAR(v_sum.x, 3.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.y, 5.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.z, 8.0f, 0.0001f);
+    ASSERT_NEAR(v_sum.x, 3.0, 0.0001);
+    ASSERT_NEAR(v_sum.y, 5.0, 0.0001);
+    ASSERT_NEAR(v_sum.z, 8.0, 0.0001);
 }
 
 TEST(TestVecPlusCorrectness, Subtest_2)
@@ -27,9 +27,9 @@ TEST(TestVecPlusCorrectness, Subtest_2)
 
     Vector3D v_sum = v1 + v2;
 
-    ASSERT_NEAR(v_sum.x, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.y, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.z, 0.0f, 0.0001f);
+    ASSERT_NEAR(v_sum.x, 0.0, 0.0001);
+    ASSERT_NEAR(v_sum.y, 0.0, 0.0001);
+    ASSERT_NEAR(v_sum.z, 0.0, 0.0001);
 }
 
 TEST(TestVecPlusCorrectness, Subtest_3)
@@ -39,9 +39,9 @@ TEST(TestVecPlusCorrectness, Subtest_3)
 
     Vector3D v_sum = v1 + v2;
 
-    ASSERT_NEAR(v_sum.x, 4.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.y, 1.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.z, 2.5f, 0.0001f);
+    ASSERT_NEAR(v_sum.x, 4.0, 0.0001);
+    ASSERT_NEAR(v_sum.y, 1.0, 0.0001);
+    ASSERT_NEAR(v_sum.z, 2.5f, 0.0001);
 }
 
 TEST(TestVecMinusCorrectness, Subtest_1)
@@ -51,9 +51,9 @@ TEST(TestVecMinusCorrectness, Subtest_1)
 
     Vector3D v_diff = v1 - v2;
 
-    ASSERT_NEAR(v_diff.x, -4.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.y, -5.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.z, -6.0f, 0.0001f);
+    ASSERT_NEAR(v_diff.x, -4.0, 0.0001);
+    ASSERT_NEAR(v_diff.y, -5.0, 0.0001);
+    ASSERT_NEAR(v_diff.z, -6.0, 0.0001);
 }
 
 TEST(TestVecMinusCorrectness, Subtest_2)
@@ -63,9 +63,9 @@ TEST(TestVecMinusCorrectness, Subtest_2)
 
     Vector3D v_diff = v1 - v2;
 
-    ASSERT_NEAR(v_diff.x, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.y, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.z, 0.0f, 0.0001f);
+    ASSERT_NEAR(v_diff.x, 0.0, 0.0001);
+    ASSERT_NEAR(v_diff.y, 0.0, 0.0001);
+    ASSERT_NEAR(v_diff.z, 0.0, 0.0001);
 }
 
 TEST(TestVecMinusCorrectness, Subtest_3)
@@ -75,111 +75,111 @@ TEST(TestVecMinusCorrectness, Subtest_3)
 
     Vector3D v_diff = v1 - v2;
 
-    ASSERT_NEAR(v_diff.x, 3.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.y, 6.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.z, -5.5f, 0.0001f);
+    ASSERT_NEAR(v_diff.x, 3.0, 0.0001);
+    ASSERT_NEAR(v_diff.y, 6.0, 0.0001);
+    ASSERT_NEAR(v_diff.z, -5.5f, 0.0001);
 }
 
 TEST(TestVecMultOnConstCorrectness, Subtest_1)
 {
-    Vector3D v(1.0f, 2.0f, 3.0f);
-    float scalar = 2.0f;
+    Vector3D v(1.0, 2.0, 3.0);
+    double scalar = 2.0;
 
     Vector3D v_scaled = v * scalar;
 
-    ASSERT_NEAR(v_scaled.x, 2.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.y, 4.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.z, 6.0f, 0.0001f);
+    ASSERT_NEAR(v_scaled.x, 2.0, 0.0001);
+    ASSERT_NEAR(v_scaled.y, 4.0, 0.0001);
+    ASSERT_NEAR(v_scaled.z, 6.0, 0.0001);
 }
 
 TEST(TestVecMultOnConstCorrectness, Subtest_2)
 {
-    Vector3D v(4.0f, 5.0f, 6.0f);
-    float scalar = 0.0f;
+    Vector3D v(4.0, 5.0, 6.0);
+    double scalar = 0.0;
 
     Vector3D v_scaled = v * scalar;
 
-    ASSERT_NEAR(v_scaled.x, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.y, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.z, 0.0f, 0.0001f);
+    ASSERT_NEAR(v_scaled.x, 0.0, 0.0001);
+    ASSERT_NEAR(v_scaled.y, 0.0, 0.0001);
+    ASSERT_NEAR(v_scaled.z, 0.0, 0.0001);
 }
 
 TEST(TestVecMultOnConstCorrectness, Subtest_3)
 {
-    Vector3D v(-1.0f, -2.0f, -3.0f);
-    float scalar = -3.0f;
+    Vector3D v(-1.0, -2.0, -3.0);
+    double scalar = -3.0;
 
     Vector3D v_scaled = v * scalar;
 
-    ASSERT_NEAR(v_scaled.x, 3.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.y, 6.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.z, 9.0f, 0.0001f);
+    ASSERT_NEAR(v_scaled.x, 3.0, 0.0001);
+    ASSERT_NEAR(v_scaled.y, 6.0, 0.0001);
+    ASSERT_NEAR(v_scaled.z, 9.0, 0.0001);
 }
 
 TEST(TestVectorCrossCorrectness, Subtest_1)
 {
-    Vector3D v1(1.0f, 0.0f, 0.0f);
-    Vector3D v2(0.0f, 1.0f, 0.0f);
+    Vector3D v1(1.0, 0.0, 0.0);
+    Vector3D v2(0.0, 1.0, 0.0);
 
     Vector3D v_cross = Cross(v1, v2);
 
-    ASSERT_NEAR(v_cross.x, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_cross.y, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_cross.z, 1.0f, 0.0001f);
+    ASSERT_NEAR(v_cross.x, 0.0, 0.0001);
+    ASSERT_NEAR(v_cross.y, 0.0, 0.0001);
+    ASSERT_NEAR(v_cross.z, 1.0, 0.0001);
 }
 
 TEST(TestVectorCrossCorrectness, Subtest_2)
 {
-    Vector3D v1(0.0f, 1.0f, 0.0f);
-    Vector3D v2(0.0f, 0.0f, 1.0f);
+    Vector3D v1(0.0, 1.0, 0.0);
+    Vector3D v2(0.0, 0.0, 1.0);
 
     Vector3D v_cross = Cross(v1, v2);
 
-    ASSERT_NEAR(v_cross.x, 1.0f, 0.0001f);
-    ASSERT_NEAR(v_cross.y, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_cross.z, 0.0f, 0.0001f);
+    ASSERT_NEAR(v_cross.x, 1.0, 0.0001);
+    ASSERT_NEAR(v_cross.y, 0.0, 0.0001);
+    ASSERT_NEAR(v_cross.z, 0.0, 0.0001);
 }
 
 TEST(TestVectorCrossCorrectness, Subtest_3)
 {
-    Vector3D v1(-1.0f, -2.0f, 0);
-    Vector3D v2(-2.0f, 0, 1.0f);
+    Vector3D v1(-1.0, -2.0, 0);
+    Vector3D v2(-2.0, 0, 1.0);
 
     Vector3D v_cross = Cross(v1, v2);
 
-    ASSERT_NEAR(v_cross.x, -2.0f, 0.0001f);
-    ASSERT_NEAR(v_cross.y,  1.0f, 0.0001f);
-    ASSERT_NEAR(v_cross.z, -4.0f, 0.0001f);
+    ASSERT_NEAR(v_cross.x, -2.0, 0.0001);
+    ASSERT_NEAR(v_cross.y,  1.0, 0.0001);
+    ASSERT_NEAR(v_cross.z, -4.0, 0.0001);
 }
 
 TEST(TestVectorDotCorrectness, Subtest_1)
 {
-    Vector3D v1(1.0f, 0.0f, 0.0f);
-    Vector3D v2(0.0f, 1.0f, 0.0f);
+    Vector3D v1(1.0, 0.0, 0.0);
+    Vector3D v2(0.0, 1.0, 0.0);
 
-    float dot_product = Dot(v1, v2);
+    double dot_product = Dot(v1, v2);
 
-    ASSERT_NEAR(dot_product, 0.0f, 0.0001f);
+    ASSERT_NEAR(dot_product, 0.0, 0.0001);
 }
 
 TEST(TestVectorDotCorrectness, Subtest_2)
 {
-    Vector3D v1(1.0f, 2.0f, 3.0f);
-    Vector3D v2(4.0f, 5.0f, 6.0f);
+    Vector3D v1(1.0, 2.0, 3.0);
+    Vector3D v2(4.0, 5.0, 6.0);
 
-    float dot_product = Dot(v1, v2);
+    double dot_product = Dot(v1, v2);
 
-    ASSERT_NEAR(dot_product, 32.0f, 0.0001f);
+    ASSERT_NEAR(dot_product, 32.0, 0.0001);
 }
 
 TEST(TestVectorDotCorrectness, Subtest_3)
 {
-    Vector3D v1(-1.0f, 4.0f, -2.0f);
-    Vector3D v2(2.0f, -3.0f, 1.0f);
+    Vector3D v1(-1.0, 4.0, -2.0);
+    Vector3D v2(2.0, -3.0, 1.0);
 
-    float dot_product = Dot(v1, v2);
+    double dot_product = Dot(v1, v2);
 
-    ASSERT_NEAR(dot_product, -16.0f, 0.0001f);
+    ASSERT_NEAR(dot_product, -16.0, 0.0001);
 }
 
 TEST(TestVectorAssignmentOperator, AssignPointToVector)
@@ -224,24 +224,24 @@ TEST(TestVectorAssignmentOperator, AssignNegativePointToVector)
 
 TEST(TestVecPlusCorrectness2D, Subtest_1)
 {
-    Vector2D v1(1.0f, 2.0f);
-    Vector2D v2(2.0f, 3.0f);
+    Vector2D v1(1.0, 2.0);
+    Vector2D v2(2.0, 3.0);
 
     Vector2D v_sum = v1 + v2;
 
-    ASSERT_NEAR(v_sum.x, 3.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.y, 5.0f, 0.0001f);
+    ASSERT_NEAR(v_sum.x, 3.0, 0.0001);
+    ASSERT_NEAR(v_sum.y, 5.0, 0.0001);
 }
 
 TEST(TestVecPlusCorrectness2D, Subtest_2)
 {
-    Vector2D v1(-1.0f, -2.0f);
-    Vector2D v2(1.0f, 2.0f);
+    Vector2D v1(-1.0, -2.0);
+    Vector2D v2(1.0, 2.0);
 
     Vector2D v_sum = v1 + v2;
 
-    ASSERT_NEAR(v_sum.x, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.y, 0.0f, 0.0001f);
+    ASSERT_NEAR(v_sum.x, 0.0, 0.0001);
+    ASSERT_NEAR(v_sum.y, 0.0, 0.0001);
 }
 
 TEST(TestVecPlusCorrectness2D, Subtest_3)
@@ -251,30 +251,30 @@ TEST(TestVecPlusCorrectness2D, Subtest_3)
 
     Vector2D v_sum = v1 + v2;
 
-    ASSERT_NEAR(v_sum.x, 4.0f, 0.0001f);
-    ASSERT_NEAR(v_sum.y, 1.0f, 0.0001f);
+    ASSERT_NEAR(v_sum.x, 4.0, 0.0001);
+    ASSERT_NEAR(v_sum.y, 1.0, 0.0001);
 }
 
 TEST(TestVecMinusCorrectness2D, Subtest_1)
 {
-    Vector2D v1(0.0f, 0.0f);
-    Vector2D v2(4.0f, 5.0f);
+    Vector2D v1(0.0, 0.0);
+    Vector2D v2(4.0, 5.0);
 
     Vector2D v_diff = v1 - v2;
 
-    ASSERT_NEAR(v_diff.x, -4.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.y, -5.0f, 0.0001f);
+    ASSERT_NEAR(v_diff.x, -4.0, 0.0001);
+    ASSERT_NEAR(v_diff.y, -5.0, 0.0001);
 }
 
 TEST(TestVecMinusCorrectness2D, Subtest_2)
 {
-    Vector2D v1(-1.0f, -2.0f);
-    Vector2D v2(-1.0f, -2.0f);
+    Vector2D v1(-1.0, -2.0);
+    Vector2D v2(-1.0, -2.0);
 
     Vector2D v_diff = v1 - v2;
 
-    ASSERT_NEAR(v_diff.x, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.y, 0.0f, 0.0001f);
+    ASSERT_NEAR(v_diff.x, 0.0, 0.0001);
+    ASSERT_NEAR(v_diff.y, 0.0, 0.0001);
 }
 
 TEST(TestVecMinusCorrectness2D, Subtest_3)
@@ -284,71 +284,71 @@ TEST(TestVecMinusCorrectness2D, Subtest_3)
 
     Vector2D v_diff = v1 - v2;
 
-    ASSERT_NEAR(v_diff.x, 3.0f, 0.0001f);
-    ASSERT_NEAR(v_diff.y, 6.0f, 0.0001f);
+    ASSERT_NEAR(v_diff.x, 3.0, 0.0001);
+    ASSERT_NEAR(v_diff.y, 6.0, 0.0001);
 }
 
 TEST(TestVecMultOnConstCorrectness2D, Subtest_1)
 {
-    Vector2D v(1.0f, 2.0f);
-    float scalar = 2.0f;
+    Vector2D v(1.0, 2.0);
+    double scalar = 2.0;
 
     Vector2D v_scaled = v * scalar;
 
-    ASSERT_NEAR(v_scaled.x, 2.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.y, 4.0f, 0.0001f);
+    ASSERT_NEAR(v_scaled.x, 2.0, 0.0001);
+    ASSERT_NEAR(v_scaled.y, 4.0, 0.0001);
 }
 
 TEST(TestVecMultOnConstCorrectness2D, Subtest_2)
 {
-    Vector2D v(4.0f, 5.0f);
-    float scalar = 0.0f;
+    Vector2D v(4.0, 5.0);
+    double scalar = 0.0;
 
     Vector2D v_scaled = v * scalar;
 
-    ASSERT_NEAR(v_scaled.x, 0.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.y, 0.0f, 0.0001f);
+    ASSERT_NEAR(v_scaled.x, 0.0, 0.0001);
+    ASSERT_NEAR(v_scaled.y, 0.0, 0.0001);
 }
 
 TEST(TestVecMultOnConstCorrectness2D, Subtest_3)
 {
-    Vector2D v(-1.0f, -2.0f);
-    float scalar = -3.0f;
+    Vector2D v(-1.0, -2.0);
+    double scalar = -3.0;
 
     Vector2D v_scaled = v * scalar;
 
-    ASSERT_NEAR(v_scaled.x, 3.0f, 0.0001f);
-    ASSERT_NEAR(v_scaled.y, 6.0f, 0.0001f);
+    ASSERT_NEAR(v_scaled.x, 3.0, 0.0001);
+    ASSERT_NEAR(v_scaled.y, 6.0, 0.0001);
 }
 
 TEST(TestVectorDotCorrectness2D, Subtest_1)
 {
-    Vector2D v1(1.0f, 0.0f);
-    Vector2D v2(0.0f, 1.0f);
+    Vector2D v1(1.0, 0.0);
+    Vector2D v2(0.0, 1.0);
 
-    float dot_product = Dot(v1, v2);
+    double dot_product = Dot(v1, v2);
 
-    ASSERT_NEAR(dot_product, 0.0f, 0.0001f);
+    ASSERT_NEAR(dot_product, 0.0, 0.0001);
 }
 
 TEST(TestVectorDotCorrectness2D, Subtest_2)
 {
-    Vector2D v1(1.0f, 2.0f);
-    Vector2D v2(3.0f, 4.0f);
+    Vector2D v1(1.0, 2.0);
+    Vector2D v2(3.0, 4.0);
 
-    float dot_product = Dot(v1, v2);
+    double dot_product = Dot(v1, v2);
 
-    ASSERT_NEAR(dot_product, 11.0f, 0.0001f);
+    ASSERT_NEAR(dot_product, 11.0, 0.0001);
 }
 
 TEST(TestVectorDotCorrectness2D, Subtest_3)
 {
-    Vector2D v1(-1.0f, 4.0f);
-    Vector2D v2(2.0f, -3.0f);
+    Vector2D v1(-1.0, 4.0);
+    Vector2D v2(2.0, -3.0);
 
-    float dot_product = Dot(v1, v2);
+    double dot_product = Dot(v1, v2);
 
-    ASSERT_NEAR(dot_product, -14.0f, 0.0001f);
+    ASSERT_NEAR(dot_product, -14.0, 0.0001);
 }
 
 TEST(TestVectorAssignmentOperator2D, AssignPointToVector)
@@ -358,55 +358,55 @@ TEST(TestVectorAssignmentOperator2D, AssignPointToVector)
 
     vector = point;
 
-    ASSERT_NEAR(vector.x, point.x, 0.0001f);
-    ASSERT_NEAR(vector.y, point.y, 0.0001f);
+    ASSERT_NEAR(vector.x, point.x, 0.0001);
+    ASSERT_NEAR(vector.y, point.y, 0.0001);
 }
 
 TEST(TestVectorAssignmentOperator2D, AssignZeroPointToVector)
 {
-    Point2D point(0.0f, 0.0f);
-    Vector2D vector(1.0f, -2.0f);
+    Point2D point(0.0, 0.0);
+    Vector2D vector(1.0, -2.0);
 
     vector = point;
 
-    ASSERT_NEAR(vector.x, point.x, 0.0001f);
-    ASSERT_NEAR(vector.y, point.y, 0.0001f);
+    ASSERT_NEAR(vector.x, point.x, 0.0001);
+    ASSERT_NEAR(vector.y, point.y, 0.0001);
 }
 
 TEST(TestVectorAssignmentOperator2D, AssignNegativePointToVector)
 {
-    Point2D point(-5.0f, -1.5f);
+    Point2D point(-5.0, -1.5f);
     Vector2D vector;
 
     vector = point;
 
-    ASSERT_NEAR(vector.x, point.x, 0.0001f);
-    ASSERT_NEAR(vector.y, point.y, 0.0001f);
+    ASSERT_NEAR(vector.x, point.x, 0.0001);
+    ASSERT_NEAR(vector.y, point.y, 0.0001);
 }
 
 TEST(TestCrossProduct, CrossOfNonZeroVectors)
 {
-    Vector2D v1(3.0f, 2.0f);
-    Vector2D v2(1.0f, 4.0f);
+    Vector2D v1(3.0, 2.0);
+    Vector2D v2(1.0, 4.0);
 
-    float result = Cross(v1, v2);
-    ASSERT_NEAR(result, 10.0f, 1e-5); // 3*4 - 2*1 = 12 - 2 = 10
+    double result = Cross(v1, v2);
+    ASSERT_NEAR(result, 10.0, 1e-5); // 3*4 - 2*1 = 12 - 2 = 10
 }
 
 TEST(TestCrossProduct, CrossOfZeroVector)
 {
-    Vector2D v1(0.0f, 0.0f);
-    Vector2D v2(1.0f, 1.0f);
-    float result = Cross(v1, v2);
-    ASSERT_NEAR(result, 0.0f, 1e-5); // 0*1 - 0*1 = 0
+    Vector2D v1(0.0, 0.0);
+    Vector2D v2(1.0, 1.0);
+    double result = Cross(v1, v2);
+    ASSERT_NEAR(result, 0.0, 1e-5); // 0*1 - 0*1 = 0
 }
 
 TEST(TestCrossProduct, CrossOfOrthogonalVectors)
 {
-    Vector2D v1(1.0f, 0.0f);
-    Vector2D v2(0.0f, 1.0f);
-    float result = Cross(v1, v2);
-    ASSERT_NEAR(result, 1.0f, 1e-5); // 1*1 - 0*0 = 1
+    Vector2D v1(1.0, 0.0);
+    Vector2D v2(0.0, 1.0);
+    double result = Cross(v1, v2);
+    ASSERT_NEAR(result, 1.0, 1e-5); // 1*1 - 0*0 = 1
 }
 
 int main(int argc, char **argv)
